% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_N.R
\name{alternative_variance_pva}
\alias{alternative_variance_pva}
\title{alternative_variance_pva}
\usage{
alternative_variance_pva(variance.b = NULL, variance.d = NULL)
}
\description{
alternative_variance_pva
}
