% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_N.R
\name{baseline_pva}
\alias{baseline_pva}
\title{baseline_pva}
\usage{
baseline_pva(number.of.years, number.of.populations, b0, d0, variance.b = 0.4,
  variance.d = 0.9, fire.prob = 0.25, fire.b, fire.d, K = NULL,
  theta = 1, d.slope = 0.0075, extinction.threshold = 0)
}
\arguments{
\item{number.of.years}{Number of years to simulate dynamics beyond the last observed count}

\item{number.of.populations}{Number simulated populations}

\item{b0}{The baseline birth rate}

\item{d0}{The baseline death rate}

\item{fire.prob}{Annual probability that a severe fire occurs}

\item{fire.b}{Birth rate in severe fire years}

\item{fire.d}{Death rate in severe fire years}

\item{K}{Carrying capacity}

\item{exitinction.threshold}{The abundance threshold used to classify quasi-extinction probability (default is 0)}
}
\description{
baseline_pva
}
