% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_assignment.R
\name{get_assignment}
\alias{get_assignment}
\title{get_assignment}
\usage{
get_assignment(assignment)
}
\arguments{
\item{assignment}{Assignment number (refer to lab activity page if you are unsure which assignment number you need to open)}
}
\description{
Open .Rmd template for weekly lab assignments
}
