% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pva.R
\name{pva}
\alias{pva}
\title{pva}
\usage{
pva(N0, number.of.years, number.of.populations, b0, variance.b = 0, d0,
  variance.d = 0, K = NULL, Allee = NULL, theta = 1,
  extinction.threshold = 0, d.slope = 0.0075)
}
\arguments{
\item{N0}{Initial population size}

\item{number.of.years}{Number of years to simulate population change}

\item{number.of.populations}{Number of simulated populations}

\item{K}{Optional carrying capacity (if K given, population growth modeled using density-dependent model)}

\item{Allee}{Optional strenth of Allee effects}

\item{theta}{Optional parameter for theta-logistic model to allow non-linear density-dependence (must be >=0)}

\item{mean.r}{Mean intrinsic population growth rate}

\item{variance.r}{Magnitude of environmental stochasticity (annual variation in `r`)}
}
\description{
Basic population viability analysis simulator
}
