% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WILD3810-package.R
\docType{package}
\name{WILD3810-package}
\alias{WILD3810}
\alias{WILD3810-package}
\title{WILD3810: WILD3810: Ecology of Plants and Animals}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: Clark Rushing \email{first.last@example.com}

}
\keyword{internal}
