% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pva.R
\name{pva}
\alias{pva}
\title{pva}
\usage{
pva(N0, number.of.years, number.of.populations = 500, r0, variance.r = 0,
  K = NULL, Allee = NULL, theta = 1, extinction.threshold = 0)
}
\arguments{
\item{N0}{Initial population size}

\item{number.of.years}{Number of years to simulate population change}

\item{number.of.populations}{Number of simulated populations}

\item{r0}{Either the mean growth rate (density-independent model) or the growth rate at N = 0 (density-dependent model)}

\item{variance.r}{Magnitude of environmental stochasticity in the growth rate (annual variation in `r`)}

\item{K}{Optional carrying capacity. If K is given, the model will assume density dependence in both b and d; if no K is given, the model will assume density-independent growth}

\item{theta}{Theta parameter in the theta-logistic growth model; control shape of density-dependence (default = 1)}

\item{extinction.threshold}{Threshold abundance used to determine a population "extinct" (default = 0)}

\item{dd.slope}{Strength of density-dependence on r}
}
\description{
Basic population viability analysis simulator
}
