% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WILD3810-package.R
\docType{package}
\name{WILD3810-package}
\alias{WILD3810}
\alias{WILD3810-package}
\title{WILD3810: WILD3810: Plant and Animal Populations}
\description{
Course package for WILD3810: Plant and Animal Populations at Utah State University. This package primarily exists to distribute data, code, and other materials in an efficient manner.
}
\author{
\strong{Maintainer}: Clark Rushing \email{clark.rushing@usu.edu}

}
\keyword{internal}
